requirements.txt
flask
flask-cors
pywebview
